@@ -1,7 +1,6 @@
 package main
 
 import (
-	"log"
 	"net/http"
 
 	"os/exec"
@@ -74,10 +73,10 @@ func hardware_power_poweroff(w http.ResponseWriter, r *http.Request) {
 		cmd := exec.Command("shutdown", "-s", "-t", "20")
 		out, err := cmd.CombinedOutput()
 		if err != nil {
-			log.Println(string(out))
+			systemWideLogger.PrintAndLog("Power", string(out), err)
 			common.SendErrorResponse(w, string(out))
 		}
-		log.Println(string(out))
+		systemWideLogger.PrintAndLog("Power", string(out), nil)
 	}
 
 	if runtime.GOOS == "linux" {
@@ -85,10 +84,10 @@ func hardware_power_poweroff(w http.ResponseWriter, r *http.Request) {
 		cmd := exec.Command("/sbin/shutdown")
 		out, err := cmd.CombinedOutput()
 		if err != nil {
-			log.Println(string(out))
+			systemWideLogger.PrintAndLog("Power", string(out), err)
 			common.SendErrorResponse(w, string(out))
 		}
-		log.Println(string(out))
+		systemWideLogger.PrintAndLog("Power", string(out), nil)
 	}
 
 	if runtime.GOOS == "darwin" {
@@ -96,10 +95,10 @@ func hardware_power_poweroff(w http.ResponseWriter, r *http.Request) {
 		cmd := exec.Command("sudo", "shutdown", "-h", "+1")
 		out, err := cmd.CombinedOutput()
 		if err != nil {
-			log.Println(string(out))
+			systemWideLogger.PrintAndLog("Power", string(out), err)
 			common.SendErrorResponse(w, string(out))
 		}
-		log.Println(string(out))
+		systemWideLogger.PrintAndLog("Power", string(out), nil)
 	}
 
 	common.SendOK(w)
@@ -141,10 +140,10 @@ func hardware_power_restart(w http.ResponseWriter, r *http.Request) {
 		cmd := exec.Command("shutdown", "-r", "-t", "20")
 		out, err := cmd.CombinedOutput()
 		if err != nil {
-			log.Println(string(out))
+			systemWideLogger.PrintAndLog("Power", string(out), err)
 			common.SendErrorResponse(w, string(out))
 		}
-		log.Println(string(out))
+		systemWideLogger.PrintAndLog("Power", string(out), nil)
 	}
 
 	if runtime.GOOS == "linux" {
@@ -152,10 +151,10 @@ func hardware_power_restart(w http.ResponseWriter, r *http.Request) {
 		cmd := exec.Command("systemctl", "reboot")
 		out, err := cmd.CombinedOutput()
 		if err != nil {
-			log.Println(string(out))
+			systemWideLogger.PrintAndLog("Power", string(out), err)
 			common.SendErrorResponse(w, string(out))
 		}
-		log.Println(string(out))
+		systemWideLogger.PrintAndLog("Power", string(out), nil)
 	}
 
 	if runtime.GOOS == "darwin" {
@@ -163,10 +162,10 @@ func hardware_power_restart(w http.ResponseWriter, r *http.Request) {
 		cmd := exec.Command("shutdown", "-r", "+1")
 		out, err := cmd.CombinedOutput()
 		if err != nil {
-			log.Println(string(out))
+			systemWideLogger.PrintAndLog("Power", string(out), err)
 			common.SendErrorResponse(w, string(out))
 		}
-		log.Println(string(out))
+		systemWideLogger.PrintAndLog("Power", string(out), nil)
 	}
 	common.SendOK(w)
 }
