@@ -13,7 +13,6 @@ import (
 	"time"
 
 	console "imuslab.com/arozos/mod/console"
-	"imuslab.com/arozos/mod/network/gzipmiddleware"
 )
 
 /*
@@ -39,19 +38,18 @@ func SetupCloseHandler() {
 
 func executeShutdownSequence() {
 	//Shutdown authAgent
-	log.Println("\r- Shutting down auth gateway")
+	systemWideLogger.PrintAndLog("System", "<!> Shutting down auth gateway", nil)
 	authAgent.Close()
 
 	//Shutdown all storage pools
-	log.Println("\r- Shutting down storage pools")
+	systemWideLogger.PrintAndLog("System", "<!> Shutting down storage pools", nil)
 	closeAllStoragePools()
 
-	//Shutdown Subservices
-	log.Println("\r- Shutting down background subservices")
-	//system_subservice_handleShutdown()
+	//Shutdown Logger
+	systemWideLogger.Close()
 
 	//Shutdown database
-	log.Println("\r- Shutting down database")
+	systemWideLogger.PrintAndLog("System", "<!> Shutting down database", nil)
 	sysdb.Close()
 
 	//Shutdown network services
@@ -59,12 +57,12 @@ func executeShutdownSequence() {
 
 	//Shutdown FTP Server
 	if ftpServer != nil {
-		log.Println("\r- Shutting down FTP Server")
+		systemWideLogger.PrintAndLog("System", "<!> Shutting down FTP Server", nil)
 		ftpServer.Close()
 	}
 
 	//Cleaning up tmp files
-	log.Println("\r- Cleaning up tmp folder")
+	systemWideLogger.PrintAndLog("System", "<!> Cleaning up tmp folder", nil)
 	os.RemoveAll(*tmp_directory)
 	//Do other things
 	os.Exit(0)
@@ -114,13 +112,18 @@ func main() {
 
 	//Initiate all the static files transfer
 	fs := http.FileServer(http.Dir("./web"))
-	if *enable_gzip {
-		//Gzip enabled. Always serve with gzip if header exists
-		http.Handle("/", gzipmiddleware.Compress(mrouter(fs)))
-	} else {
-		//Normal file server without gzip
-		http.Handle("/", mrouter(fs))
-	}
+	/*
+		if *enable_gzip {
+			//Gzip enabled. Always serve with gzip if header exists
+			//http.Handle("/", gzipmiddleware.Compress(mrouter(fs)))
+			http.Handle("/", mrouter(fs))
+		} else {
+			//Normal file server without gzip
+			http.Handle("/", mrouter(fs))
+		}
+	*/
+	//Updates 2022-09-06: Gzip handler moved inside the master router
+	http.Handle("/", mrouter(fs))
 
 	//Set database read write to ReadOnly after startup if demo mode
 	if *demo_mode {
