@@ -15,6 +15,7 @@ import (
 
 	"imuslab.com/arozos/mod/common"
 	fs "imuslab.com/arozos/mod/filesystem"
+	"imuslab.com/arozos/mod/network/gzipmiddleware"
 )
 
 func mrouter(h http.Handler) http.Handler {
@@ -68,6 +69,8 @@ func mrouter(h http.Handler) http.Handler {
 			WebDavHandler.HandleRequest(w, r)
 		} else if len(r.URL.Path) >= len("/share") && r.URL.Path[:6] == "/share" {
 			shareManager.HandleShareAccess(w, r)
+		} else if len(r.URL.Path) >= len("/api/remote") && r.URL.Path[:11] == "/api/remote" {
+			AGIGateway.ExtAPIHandler(w, r)
 		} else if r.URL.Path == "/" && authAgent.CheckAuth(r) {
 			//Use logged in and request the index. Serve the user's interface module
 			w.Header().Set("Cache-Control", "no-cache, no-store, no-transform, must-revalidate, private, max-age=0")
@@ -118,7 +121,6 @@ func mrouter(h http.Handler) http.Handler {
 				//Enable subservice access
 				//Check if this path is reverse proxy path. If yes, serve with proxyserver
 				isRP, proxy, rewriteURL, subserviceObject := ssRouter.CheckIfReverseProxyPath(r)
-
 				if isRP {
 					//Check user permission on that module
 					ssRouter.HandleRoutingRequest(w, r, proxy, subserviceObject, rewriteURL)
@@ -144,8 +146,7 @@ func mrouter(h http.Handler) http.Handler {
 				errorHandleNotFound(w, r)
 				return
 			}
-			h.ServeHTTP(w, r)
-
+			routerStaticContentServer(h, w, r)
 		} else {
 			//User not logged in. Check if the path end with public/. If yes, allow public access
 			if !fs.FileExists(filepath.Join("./web", r.URL.Path)) {
@@ -153,10 +154,10 @@ func mrouter(h http.Handler) http.Handler {
 				errorHandleNotFound(w, r)
 			} else if r.URL.Path[len(r.URL.Path)-1:] != "/" && filepath.Base(filepath.Dir(r.URL.Path)) == "public" {
 				//This file path end with public/. Allow public access
-				h.ServeHTTP(w, r)
-			} else if *allow_homepage == true && len(r.URL.Path) >= 5 && r.URL.Path[:5] == "/www/" {
+				routerStaticContentServer(h, w, r)
+			} else if *allow_homepage && len(r.URL.Path) >= 5 && r.URL.Path[:5] == "/www/" {
 				//Handle public home serving if homepage mode is enabled
-				h.ServeHTTP(w, r)
+				routerStaticContentServer(h, w, r)
 			} else {
 				//Other paths
 				//Rediect to login page
@@ -167,4 +168,12 @@ func mrouter(h http.Handler) http.Handler {
 		}
 
 	})
+}
+
+func routerStaticContentServer(h http.Handler, w http.ResponseWriter, r *http.Request) {
+	if *enable_gzip {
+		gzipmiddleware.Compress(h).ServeHTTP(w, r)
+	} else {
+		h.ServeHTTP(w, r)
+	}
 }
