@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/robertkrimen/otto"
+	"imuslab.com/arozos/mod/filesystem"
 	user "imuslab.com/arozos/mod/user"
 )
 
@@ -15,7 +16,7 @@ func (g *Gateway) ShareLibRegister() {
 	}
 }
 
-func (g *Gateway) injectShareFunctions(vm *otto.Otto, u *user.User) {
+func (g *Gateway) injectShareFunctions(vm *otto.Otto, u *user.User, scriptFsh *filesystem.FileSystemHandler, scriptPath string) {
 	vm.Set("_share_file", func(call otto.FunctionCall) otto.Value {
 		//Get the vpath of file to share
 		vpath, err := call.Argument(0).ToString()
@@ -30,19 +31,9 @@ func (g *Gateway) injectShareFunctions(vm *otto.Otto, u *user.User) {
 			timeout = 0
 		}
 
-		//Resolve the file path
-		rpath, err := u.VirtualPathToRealPath(vpath)
-		if err != nil {
-			return otto.New().MakeCustomError("Path resolve failed", "Unable to resolve virtual path: "+err.Error())
-		}
-
-		//Check if file exists
-		if !fileExists(rpath) {
-			return otto.New().MakeCustomError("File not exists", "Share file vpath not exists")
-		}
-
 		//Create a share object for this request
-		shareID, err := g.Option.ShareManager.CreateNewShare(u, vpath)
+		vpathSourceFsh := u.GetRootFSHFromVpathInUserScope(vpath)
+		shareID, err := g.Option.ShareManager.CreateNewShare(u, vpathSourceFsh, vpath)
 		if err != nil {
 			log.Println("[AGI] Create Share Failed: " + err.Error())
 			return otto.New().MakeCustomError("Share failed", err.Error())
@@ -51,7 +42,7 @@ func (g *Gateway) injectShareFunctions(vm *otto.Otto, u *user.User) {
 		if timeout > 0 {
 			go func(timeout int) {
 				time.Sleep(time.Duration(timeout) * time.Second)
-				g.Option.ShareManager.RemoveShareByUUID(shareID.UUID)
+				g.Option.ShareManager.RemoveShareByUUID(u, shareID.UUID)
 				log.Println("[AGI] Share auto-removed: " + shareID.UUID)
 			}(int(timeout))
 		}
@@ -65,7 +56,7 @@ func (g *Gateway) injectShareFunctions(vm *otto.Otto, u *user.User) {
 		if err != nil {
 			return otto.New().MakeCustomError("Failed to remove share", "No share UUID given")
 		}
-		err = g.Option.ShareManager.RemoveShareByUUID(shareUUID)
+		err = g.Option.ShareManager.RemoveShareByUUID(u, shareUUID)
 		if err != nil {
 			log.Println("[AGI] Share remove failed: " + err.Error())
 			return otto.New().MakeCustomError("Failed to remove share", err.Error())
@@ -81,13 +72,7 @@ func (g *Gateway) injectShareFunctions(vm *otto.Otto, u *user.User) {
 			return otto.NullValue()
 		}
 
-		//Resolve the vpath to realpath
-		rpath, err := u.VirtualPathToRealPath(vpath)
-		if err != nil {
-			log.Println("[AGI] Failed to get share UUID: Unabel to resovle")
-			return otto.NullValue()
-		}
-		shareObject := g.Option.ShareManager.GetShareObjectFromRealPath(rpath)
+		shareObject := g.Option.ShareManager.GetShareObjectFromUserAndVpath(u, vpath)
 		if shareObject == nil {
 			log.Println("[AGI] Failed to get share UUID: File not shared")
 			return otto.NullValue()
@@ -129,12 +114,7 @@ func (g *Gateway) injectShareFunctions(vm *otto.Otto, u *user.User) {
 			return otto.New().MakeCustomError("Failed to check share exists", "No filepath given")
 		}
 
-		rpath, err := u.VirtualPathToRealPath(vpath)
-		if err != nil {
-			return otto.New().MakeCustomError("Filepath resolve failed", err.Error())
-		}
-
-		isShared := g.Option.ShareManager.FileIsShared(rpath)
+		isShared := g.Option.ShareManager.FileIsShared(u, vpath)
 		r, _ := otto.ToValue(isShared)
 		return r
 	})
