@@ -9,6 +9,7 @@ import (
 	"sort"
 	"strconv"
 
+	"imuslab.com/arozos/mod/filesystem"
 	user "imuslab.com/arozos/mod/user"
 )
 
@@ -47,14 +48,16 @@ func (s *LargeFileScanner) HandleLargeFileList(w http.ResponseWriter, r *http.Re
 	type FileObject struct {
 		Filename string
 		Filepath string
-		realpath string
 		Size     int64
 		IsOwner  bool
+
+		realpath string
+		thisfsh  *filesystem.FileSystemHandler
 	}
 	//Walk all filesystem handlers and buffer all files and their sizes
 	fileList := []*FileObject{}
 	for _, fsh := range fsHandlers {
-		filepath.Walk(fsh.Path, func(path string, info os.FileInfo, err error) error {
+		fsh.FileSystemAbstraction.Walk(fsh.Path, func(path string, info os.FileInfo, err error) error {
 			if info == nil || err != nil {
 				//Disk IO Error
 				return errors.New("Disk IO Error: " + err.Error())
@@ -66,11 +69,12 @@ func (s *LargeFileScanner) HandleLargeFileList(w http.ResponseWriter, r *http.Re
 
 			//Push the current file into the filelist
 			if info.Size() > 0 {
-				vpath, _ := userinfo.RealPathToVirtualPath(path)
+				vpath, _ := fsh.FileSystemAbstraction.RealPathToVirtualPath(path, userinfo.Username)
 				fileList = append(fileList, &FileObject{
 					Filename: filepath.Base(path),
 					Filepath: vpath,
 					realpath: path,
+					thisfsh:  fsh,
 					Size:     info.Size(),
 					IsOwner:  false,
 				})
@@ -99,7 +103,7 @@ func (s *LargeFileScanner) HandleLargeFileList(w http.ResponseWriter, r *http.Re
 
 	//Only check ownership of those requested
 	for _, file := range fileList[:limitInt] {
-		if userinfo.IsOwnerOfFile(file.realpath) {
+		if userinfo.IsOwnerOfFile(file.thisfsh, file.Filepath) {
 			file.IsOwner = true
 		} else {
 			file.IsOwner = false
