@@ -7,11 +7,10 @@ import (
 	"image/jpeg"
 	"image/png"
 	"log"
-	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/nfnt/resize"
+	"imuslab.com/arozos/mod/filesystem"
 )
 
 /*
@@ -19,20 +18,24 @@ import (
 	The preview is generated by overlapping 2 - 3 layers of images
 */
 
-func generateThumbnailForFolder(cacheFolder string, file string, generateOnly bool) (string, error) {
+func generateThumbnailForFolder(fsh *filesystem.FileSystemHandler, cacheFolder string, file string, generateOnly bool) (string, error) {
+	if fsh.RequireBuffer {
+		return "", nil
+	}
+	fshAbs := fsh.FileSystemAbstraction
 	//Check if this folder has cache image folder
 	cacheFolderInsideThisFolder := filepath.Join(file, "/.metadata/.cache")
-	if !fileExists(cacheFolderInsideThisFolder) {
+	if !fshAbs.FileExists(cacheFolderInsideThisFolder) {
 		//This folder do not have a cache folder
 		return "", errors.New("No previewable files")
 	}
 
 	//Load the base template
-	if !fileExists("web/img/system/folder-preview.png") {
+	if !fshAbs.FileExists("web/img/system/folder-preview.png") {
 		//Missing system files. Skip rendering
 		return "", errors.New("Missing system template image file")
 	}
-	image1, err := os.Open("web/img/system/folder-preview.png")
+	image1, err := fshAbs.Open("web/img/system/folder-preview.png")
 	if err != nil {
 		return "", err
 	}
@@ -49,12 +52,12 @@ func generateThumbnailForFolder(cacheFolder string, file string, generateOnly bo
 	draw.Draw(resultThumbnail, b, baseTemplate, image.ZP, draw.Over)
 
 	//Get cached file inside this folder, only include jpg (non folder)
-	contentCache, _ := wGlob(filepath.Join(cacheFolderInsideThisFolder, "/*.jpg"))
+	contentCache, _ := fshAbs.Glob(filepath.Join(cacheFolderInsideThisFolder, "/*.jpg"))
 
 	//Check if there are more than 1 file inside this folder that is cached
 	if len(contentCache) > 1 {
 		//More than 1 files. Render the image at the back
-		image2, err := os.Open(contentCache[1])
+		image2, err := fshAbs.Open(contentCache[1])
 		if err != nil {
 			return "", err
 		}
@@ -72,7 +75,7 @@ func generateThumbnailForFolder(cacheFolder string, file string, generateOnly bo
 	}
 
 	//Render the top image
-	image3, err := os.Open(contentCache[0])
+	image3, err := fshAbs.Open(contentCache[0])
 	if err != nil {
 		return "", errors.New("failed to open: " + err.Error())
 	}
@@ -81,7 +84,7 @@ func generateThumbnailForFolder(cacheFolder string, file string, generateOnly bo
 	if err != nil {
 		//Fail to decode the image. Try to remove the damaged iamge file
 		image3.Close()
-		os.Remove(contentCache[0])
+		fshAbs.Remove(contentCache[0])
 		log.Println("Failed to decode cahce image for: " + contentCache[0] + ". Removing thumbnail cache")
 		return "", errors.New("failed to decode: " + err.Error())
 	}
@@ -91,42 +94,13 @@ func generateThumbnailForFolder(cacheFolder string, file string, generateOnly bo
 	resizedTopImage := resize.Resize(260, 260, topImage, resize.Lanczos3)
 	draw.Draw(resultThumbnail, resizedTopImage.Bounds().Add(topImageOffset), resizedTopImage, image.ZP, draw.Over)
 
-	outfile, err := os.Create(filepath.Join(cacheFolder, filepath.Base(file)+".png"))
+	outfile, err := fshAbs.Create(filepath.Join(cacheFolder, filepath.Base(file)+".png"))
 	if err != nil {
 		log.Fatalf("failed to create: %s", err)
 	}
 	png.Encode(outfile, resultThumbnail)
 	outfile.Close()
 
-	ctx, err := getImageAsBase64(cacheFolder + filepath.Base(file) + ".png")
+	ctx, err := getImageAsBase64(fsh, cacheFolder+filepath.Base(file)+".png")
 	return ctx, err
-}
-
-//Special glob function to handle file name containing wildcard characters
-func wGlob(path string) ([]string, error) {
-	files, err := filepath.Glob(path)
-	if err != nil {
-		return []string{}, err
-	}
-
-	if strings.Contains(path, "[") == true || strings.Contains(path, "]") == true {
-		if len(files) == 0 {
-			//Handle reverse check. Replace all [ and ] with ?
-			newSearchPath := strings.ReplaceAll(path, "[", "?")
-			newSearchPath = strings.ReplaceAll(newSearchPath, "]", "?")
-			//Scan with all the similar structure except [ and ]
-			tmpFilelist, _ := filepath.Glob(newSearchPath)
-			for _, file := range tmpFilelist {
-				file = filepath.ToSlash(file)
-				if strings.Contains(file, filepath.ToSlash(filepath.Dir(path))) {
-					files = append(files, file)
-				}
-			}
-		}
-	}
-	//Convert all filepaths to slash
-	for i := 0; i < len(files); i++ {
-		files[i] = filepath.ToSlash(files[i])
-	}
-	return files, nil
 }
