@@ -19,6 +19,9 @@ import (
 	"sync"
 	"time"
 
+	"imuslab.com/arozos/mod/filesystem"
+	"imuslab.com/arozos/mod/filesystem/hidden"
+	"imuslab.com/arozos/mod/filesystem/metadata"
 	"imuslab.com/arozos/mod/network/webdav"
 	"imuslab.com/arozos/mod/user"
 )
@@ -253,16 +256,25 @@ func (s *Server) HandleRequest(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	//Try to resolve the realpath of the vroot
-	realRoot, err := userinfo.VirtualPathToRealPath(reqRoot + ":/")
+	fsh, err := userinfo.GetFileSystemHandlerFromVirtualPath(reqRoot + ":/")
 	if err != nil {
-		log.Println(err.Error())
+		log.Println("[WebDAV] Failed to load File System Handler from request root: ", reqRoot+":/", err.Error())
 		http.Error(w, "Invalid ", http.StatusUnauthorized)
 		return
 	}
 
+	//Try to resolve the realpath of the vroot
+	/*
+		realRoot, err := userinfo.VirtualPathToRealPath(reqRoot + ":/")
+		if err != nil {
+			log.Println(err.Error())
+			http.Error(w, "Invalid ", http.StatusUnauthorized)
+			return
+		}
+	*/
+
 	//Ok. Check if the file server of this root already exists
-	fs := s.getFsFromRealRoot(realRoot, filepath.ToSlash(filepath.Join(s.prefix, reqRoot)))
+	fs := s.getFsFromRealRoot(fsh, userinfo.Username, filepath.ToSlash(filepath.Join(s.prefix, reqRoot)))
 
 	//Serve the content
 	fs.ServeHTTP(w, r)
@@ -287,21 +299,45 @@ func (s *Server) serveReadOnlyWebDav(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-func (s *Server) getFsFromRealRoot(realRoot string, prefix string) *webdav.Handler {
-	tfs, ok := s.filesystems.Load(realRoot)
-	if !ok {
-		//This file system handle hasn't been created. Create it now
-		fs := &webdav.Handler{
-			Prefix:     prefix,
-			FileSystem: webdav.Dir(realRoot),
-			LockSystem: webdav.NewMemLS(),
-		}
-
-		//Store the file system handler
-		s.filesystems.Store(realRoot, fs)
-
-		return fs
-	} else {
-		return tfs.(*webdav.Handler)
+func (s *Server) getFsFromRealRoot(fsh *filesystem.FileSystemHandler, username string, prefix string) *webdav.Handler {
+	//Create a webdav adapter from the fsh
+	fshadapter := NewFshWebDAVAdapter(fsh, username)
+	fs := &webdav.Handler{
+		Prefix:     prefix,
+		FileSystem: fshadapter,
+		LockSystem: webdav.NewMemLS(),
 	}
+
+	//Create event listener for the path request
+	fs.RequestEventListener = func(path string) {
+		//Generate thumbnail in the background if listed
+		vpath, _ := fsh.FileSystemAbstraction.RealPathToVirtualPath(path, username)
+		go func() {
+			isHidden, _ := hidden.IsHidden(vpath, false)
+			if !isHidden {
+				metadata.NewRenderHandler().BuildCacheForFolder(fsh, vpath, username)
+			}
+
+		}()
+	}
+
+	return fs
+	/*
+		tfs, ok := s.filesystems.Load(realRoot)
+		if !ok {
+			//This file system handle hasn't been created. Create it now
+			fs := &webdav.Handler{
+				Prefix:     prefix,
+				FileSystem: webdav.Dir(realRoot),
+				LockSystem: webdav.NewMemLS(),
+			}
+
+			//Store the file system handler
+			s.filesystems.Store(realRoot, fs)
+
+			return fs
+		} else {
+			return tfs.(*webdav.Handler)
+		}
+	*/
 }
